Calculates employment & sales growth
This script is designed to determine the growth in employment and sales
of firms in the year before their most recent relocation and from the
year that they opened to the year of their most recent relocation.

If a firm has never relocated, growth in employment and sales is
calculated from using the last year that the establishment was open.